requirements.txt
gradio
chromadb
faiss-cpu
sentence-transformers
PyYAML
requests
python-dotenv
pypdf
PyMuPDF
google-genai


